+++ src/pages/index/components/layouts/PageHeader.vue
@@ -53,7 +53,7 @@
 </template>
 
 <script>
-  import { isMacLike } from '../../../../utils/courseUtils';
+  import { isMacLike } from '../../../../utils/course';
 
 
   export default {
